[fix] 무음모드와 관계없이 소리 나오도록 수정
